% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ds.R
\name{compute_ds}
\alias{compute_ds}
\title{Computes Descriptive Statistics}
\usage{
compute_ds(data, var1, var2)
}
\arguments{
\item{data}{input dataset}

\item{var1}{name of variable 1 (unquoted)}

\item{var2}{name of variable 2 (unquoted)}
}
\value{
a data frame with descriptive statistics including average, min-max and standard dev
}
\description{
Computes Descriptive Statistics
}
\examples{
compute_ds(data = faithful,var1 = eruptions, var2 = waiting)

}
