% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ds_another_way.R
\name{compute_ds_another_way}
\alias{compute_ds_another_way}
\title{Computes Descriptive Statistics}
\usage{
compute_ds_another_way(data, var1, var2)
}
\arguments{
\item{data}{input data set}

\item{var1}{name of variable 1 (in quotes)}

\item{var2}{name of variable 2 (in quotes)}
}
\value{
a data frame containing descriptive statistics
}
\description{
Computes Descriptive Statistics
}
\examples{
compute_ds_another_way(data = faithful,var1 = 'eruptions', var2 = 'waiting')

}
