% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{create_package}
\alias{create_package}
\title{Create a package, optionally initialising useful stuff}
\usage{
create_package(
  path,
  use_renv = TRUE,
  use_git = TRUE,
  use_testthat = TRUE,
  use_makefile = TRUE,
  readme = c("md", "Rmd")
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}
}
\description{
Create a package, optionally initialising useful stuff
}
