% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_corr.R
\name{compute_corr}
\alias{compute_corr}
\title{Compute a tidy correlation}
\usage{
compute_corr(data, var1, var2)
}
\arguments{
\item{data}{input dataframe}

\item{var1}{name of first column}

\item{var2}{name of second column}
}
\value{
a dataframe with Pearson correlation and P-value
}
\description{
Compute a tidy correlation
}
\examples{
compute_corr(mtcars, mpg, wt)
}
